Make Repository.Query inner class, commented all of the affected code
